langgraph-dev/jobs-scraper/pyproject.toml
[project]
name = "jobs-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.5.1",
    "pillow>=11.3.0",
    "playwright>=1.53.0",
]
